+++ test/test_optimization.py
@@ -6,6 +6,8 @@ Created on Thu Mar 18 08:49:08 2021.
 @author: fabian
 """
 
+import platform
+
 import numpy as np
 import pandas as pd
 import pytest
@@ -24,6 +26,12 @@ if "gurobi" in available_solvers:
     params.append(("gurobi", "direct"))
 if "highs" in available_solvers:
     params.append(("highs", "direct"))
+
+feasible_quadratic_solvers = quadratic_solvers
+# There seems to be a bug in scipopt with quadratic models on windows, see
+# https://github.com/PyPSA/linopy/actions/runs/7615240686/job/20739454099?pr=78
+if platform.system() == "Windows":
+    feasible_quadratic_solvers.discard("scipopt")
 
 
 @pytest.fixture
@@ -475,7 +483,7 @@ def test_milp_model_r(milp_model_r, solver, io_api):
 
 @pytest.mark.parametrize("solver,io_api", params)
 def test_quadratic_model(quadratic_model, solver, io_api):
-    if solver in quadratic_solvers:
+    if solver in feasible_quadratic_solvers:
         status, condition = quadratic_model.solve(solver, io_api=io_api)
         assert condition == "optimal"
         assert (quadratic_model.solution.x.round(3) == 0).all()
@@ -488,7 +496,7 @@ def test_quadratic_model(quadratic_model, solver, io_api):
 
 @pytest.mark.parametrize("solver,io_api", params)
 def test_quadratic_model_cross_terms(quadratic_model_cross_terms, solver, io_api):
-    if solver in quadratic_solvers:
+    if solver in feasible_quadratic_solvers:
         status, condition = quadratic_model_cross_terms.solve(solver, io_api=io_api)
         assert condition == "optimal"
         assert (quadratic_model_cross_terms.solution.x.round(3) == 1.5).all()
@@ -502,7 +510,7 @@ def test_quadratic_model_cross_terms(quadratic_model_cross_terms, solver, io_api
 @pytest.mark.parametrize("solver,io_api", params)
 def test_quadratic_model_wo_constraint(quadratic_model, solver, io_api):
     quadratic_model.constraints.remove("con0")
-    if solver in quadratic_solvers:
+    if solver in feasible_quadratic_solvers:
         status, condition = quadratic_model.solve(solver, io_api=io_api)
         assert condition == "optimal"
         assert (quadratic_model.solution.x.round(3) == 0).all()
@@ -515,7 +523,7 @@ def test_quadratic_model_wo_constraint(quadratic_model, solver, io_api):
 @pytest.mark.parametrize("solver,io_api", params)
 def test_quadratic_model_unbounded(quadratic_model, solver, io_api):
     quadratic_model.objective = -quadratic_model.objective
-    if solver in quadratic_solvers:
+    if solver in feasible_quadratic_solvers:
         status, condition = quadratic_model.solve(solver, io_api=io_api)
         # particular case for scip:
         if solver == "scip":