+++ doc/release_notes.rst
@@ -7,6 +7,7 @@ Upcoming Release
 
 * Linopy now supports the SCIP solver via the `pyscipopt` package. The solver is available for free for general use. See the `SCIP website <https://www.scipopt.org/>` for more information.
 * Linopy was refactored to use the new xarray API (>=2024.01) without the deprecation warnings.
+* The set "quadratic_solvers" now only contains quadratic solvers which are installed and available to the user.
 
 Version 0.3.2
 -------------